Aggiunta post paziente, classe Checker per la validazione dei dati passati nel json, classe Regex con le espressioni regolari per controllare alcuni parametri con formati specifici

+++ src/main/java/server/database2/PatientDB.java
@@ -72,7 +72,7 @@ public class PatientDB {
             PreparedStatement st = conn.prepareStatement(sql);
             st.setString(1, String.valueOf(map.get("name")));
             st.setString(2, String.valueOf(map.get("surname")));
-            st.setInt(3, (Integer) map.get("age"));
+            st.setInt(3, ((Double) map.get("age")).intValue());
             st.setString(4, String.valueOf(map.get("phone")));
             st.setString(5, String.valueOf(map.get("address_home")));
             st.setString(6, String.valueOf(map.get("address_hospital")));
